backend/requirements.txt
fastapi==0.112.1
uvicorn[standard]==0.30.6
pydantic==2.8.2
pandas==2.2.2
python-dateutil==2.9.0.post0
jinja2==3.1.4
websockets==12.0
